- 'x64' platform added to configurations; - manifest file excluded from usage - additional manifest dependencies in linker are used now; - some files are now survive x64 compilation.
git-svn-id: http://micro-profiler.googlecode.com/svn/trunk@485 1676f2fc-95af-a39c-d699-6f41337c5309

+++ micro-profiler/collector/analyzer.cpp
@@ -29,7 +29,7 @@ namespace micro_profiler
 	void analyzer::clear()
 	{	_statistics.clear();	}
 
-	void analyzer::accept_calls(unsigned int threadid, const call_record *calls, unsigned int count)
+	void analyzer::accept_calls(unsigned int threadid, const call_record *calls, size_t count)
 	{
 		stacks_container::iterator i = _stacks.find(threadid);
